backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.1.0
django-extra-fields==3.0.2
psycopg2-binary==2.9.9
django-filter==23.5
python-dotenv==1.0.1
gunicorn==20.1.0
Pillow==9.0.0
tqdm==4.67.1